+++ pkg/providers/property/testdata/TestDataEdgeHostnames/edgehostnames_with_options.tf
@@ -1,0 +1,9 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_edge_hostnames" "test" {
+  contract_id = "ctr_123"
+  group_id    = "grp_123"
+  options     = ["mapDetails"]
+}